Reorganize spec helpers into spec/support
